+++ src/Component/Http/Message.php
@@ -9,7 +9,7 @@
 	abstract class Message implements MessageInterface {
 
 		/** @var string[] */
-		const HTTP_METHOD = [ 'GET', 'HEAD', 'POST', 'PUT', 'DELETE', 'CONNECT', 'OPTIONS', 'TRACE', 'PATCH' ];
+		const HTTP_METHODS = [ 'GET', 'HEAD', 'POST', 'PUT', 'DELETE', 'CONNECT', 'OPTIONS', 'TRACE', 'PATCH' ];
 
 		/** @var string[] */
 		const HTTP_STATUS = [